rename file from Permalink to Share
